+++ programs/dibio.rs
@@ -470,9 +470,8 @@ pub unsafe fn DiB_trainFromFiles(
         };
     }
     srcBuffer = malloc(loadedSize.wrapping_add(NOISELENGTH as size_t));
-    sampleSizes =
-        malloc((fs.nbSamples as size_t).wrapping_mul(::core::mem::size_of::<size_t>() as size_t))
-            as *mut size_t;
+    sampleSizes = malloc((fs.nbSamples as size_t).wrapping_mul(::core::mem::size_of::<size_t>()))
+        as *mut size_t;
     if fs.nbSamples != 0 && sampleSizes.is_null() || srcBuffer.is_null() || dictBuffer.is_null() {
         fprintf(
             stderr,
